fix: moved StickyNote component moved StickyNote to component folder as it was failing to be found on vercel deployment
