+++ src/app/components/StickyNote.tsx
